refactor: split tag id model
