+++ api/context/article/domain/model/tag.go
@@ -12,28 +12,6 @@ var (
 	patternTagName = regexp.MustCompile("^[\u4e00-\u9fa5ぁ-んァ-ンa-zA-Z0-9-_ ]$")
 )
 
-// TagID is a value object to identify tag
-type TagID struct {
-	model.ValueObject
-	articleID *ArticleID
-	name      string
-}
-
-// ArticleID returns the linked article's id
-func (id TagID) ArticleID() *ArticleID {
-	return id.articleID
-}
-
-// Name returns the tag's name
-func (id TagID) Name() string {
-	return id.name
-}
-
-// Equals compares tag id with another
-func (id TagID) Equals(another *TagID) bool {
-	return (id.ArticleID() == another.ArticleID()) && (id.Name() == another.Name())
-}
-
 // Tag is the tag model
 type Tag struct {
 	model.Entity
